+++ tests/PowerUtils.Text.Benchmarks/Program.cs
@@ -1,0 +1,3 @@
+﻿using PowerUtils.Text.Benchmarks.NetworkExtensionsBenchmarks;
+
+BenchmarkRunner.Run<CombineURLBenchmarks>();